allow changes to the boxfile recreate services
https://github.com/nanobox-io/nanobox-server/issues/17
fixes #17

This makes it so if you change a service in the boxfile those
changes are pushed through to the containers. This falls
inline more accurately with nanobox desktops local policy that
the boxfile is king.

It also addresses a issue where we had written a 'combinedBox'
function in multiple places to facilitate different use cases
but this left the system disjointed and anytime we needed to
modify the way boxfiles were read we had to make changes in
three places.